decathlon_agent/studio/requirements.txt
langchain
langchain_openai
tiktoken
tenacity
python-dotenv
openai
pydantic